モデルの修正
モデルをスプライトと名前を変更（スクラッチの呼び方に合わせた）
ブロック名「スプライトを作成（spawn sprite）」に変更（s2ar.js）
ファイル名の変更（Models.swift => Sprites.swift）
スプライトのコスチュームを変更できるようにした。a, b より選択する。
スプライト（8x8x8サイズ）の中心で正しく回転するように、arGameファンクション（ViewController.swift）を修正
README_JA にコスチュームを変更できることを追記。ブロックの写真を変更。
