+++ extension/s2ar.js
@@ -99,8 +99,8 @@
         socket.emit("from_client", JSON.stringify({roomId: roomId, command: command}));
       }
 
-      ext.ar_game = function(x, y, z, model, color, direction) {
-        let command = "ar_game:" + x + ":" + y + ":" + z + ":" + model + ":" + color + ":" + direction;
+      ext.ar_game = function(x, y, z, sprite, costume, color, direction) {
+        let command = "ar_game:" + x + ":" + y + ":" + z + ":" + sprite  + ":" + costume + ":" + color + ":" + direction;
         socket.emit("from_client", JSON.stringify({roomId: roomId, command: command}));
       }
 
@@ -168,7 +168,7 @@
           map: '地図を作成。地図データを %s 、幅を %n 、高さを %n 、拡大倍率を %n 、（低地の色を r1: %n g1: %n b1: %n ）、（高地の色を r2: %n g2: %n b2: %n ）、上方向へ %n',
           pin: 'ピンを立てる。位置データを %s 、幅を %n 、高さを %n 、拡大倍率を %n 、左上緯度を %n 、左上経度を %n 、右下緯度を %n 、右下経度を %n 横に %n ずらす',
           molecular_structure: '分子構造モデルを作成。x座標を %n 、y座標を %n 、z座標を %n 、拡大倍率を %n 、MLDファイル %s',
-          ar_game: 'キャラクターを作成。X座標: %n Y座標: %n Z座標: %n モデル: %m.model 色: %m.color 向き: %m.direction',
+          ar_game: 'スプライトを作成。X座標: %n Y座標: %n Z座標: %n スプライト: %m.sprite コスチューム： %m.costume 色: %m.color 向き: %m.direction',
           set_color: 'ブロックの色を変える。r: %n g: %n b: %n',
           set_alpha: 'ブロックの透明度を変える。alpha: %n',
           change_layer: 'ARのレイヤを変える。レイヤ: %m.layer',
@@ -196,7 +196,7 @@
           map: 'draw map from csv: %s width: %n height: %n magnification: %n (lowland r1: %n g1: %n b1: %n ) (highland r2: %n g2: %n b2: %n ) upward: %n',
           pin: 'stand pins at position: %s width: %n height: %n magnification: %n up-left (latitude: %n longitude: %n ) down-right (latitude: %n longitude: %n ) shift %n',
           molecular_structure: 'molecular structure at x: %n y: %n z: %n magnification: %n mld file: %s',
-          ar_game: 'spawn character at x: %n y: %n z: %n model: %m.model color: %m.color direction: %m.direction',
+          ar_game: 'spawn sprite at x: %n y: %n z: %n sprite: %m.sprite costume: %m.costume color: %m.color direction: %m.direction',
           set_color: 'set color to r: %n g: %n b: %n',
           set_alpha: 'set transparency to alpha: %n',
           change_layer: 'change AR layer: %m.layer',
@@ -227,7 +227,7 @@
           [' ', locale[lang].map, 'map', 'map_data.csv', 257, 257, 100, 0, 255, 0, 124, 96, 53, 0],
           [' ', locale[lang].pin, 'pin', 'potision_data.csv', 257, 257, 2, 46.852, 126.738, 29.148, 149.238, 0],
           [' ', locale[lang].molecular_structure, 'molecular_structure', 0, 10, 0, 10, 'methane.mld'],
-          [' ', locale[lang].ar_game, 'ar_game', 0, 0, 0, 'spaceship', 'red', '+y'],
+          [' ', locale[lang].ar_game, 'ar_game', 0, 0, 0, 'spaceship', 'a', 'red', '+y'],
           [' ', locale[lang].set_color, 'set_color', 255, 255, 255],
           [' ', locale[lang].set_alpha, 'set_alpha', 1.0],
           [' ', locale[lang].change_layer, 'change_layer', '1'],
@@ -243,7 +243,8 @@
           layer: ['1', '2', '3'],
           shape: ['cube', 'sphere', 'cylinder', 'cone', 'pyramid'],
           material: ['none', 'aluminum', 'asphalt', 'brass', 'brick', 'cedar', 'craft', 'grass', 'maple', 'marble01', 'marble02', 'punching metal', 'stainless steel', 'stone01', 'stone02', 'terra cotta', 'earth'],
-          model: ['spaceship', 'invader', 'human', 'cat', 'dinosaur', 'frog', 'tree', 'car', 'airplane', 'sword', 'fire', 'explosion', '1x1x1', '2x2x2', '4x4x4', '8x8x8'],
+          sprite: ['spaceship', 'invader', 'human', 'cat', 'dinosaur', 'frog', 'tree', 'car', 'airplane', 'sword', 'fire', 'explosion', '1x1x1', '2x2x2', '4x4x4', '8x8x8'],
+          costume: ['a', 'b'],
           color: ['red', 'green', 'blue', 'yellow', 'cyan', 'magenta', 'white', 'black'],
           direction: ['+y', '+x', '+z', '-y', '-x', '-z'],
         }
