--- ios_app/s2ar/models.swift
@@ -1,578 +1,0 @@
-//
-//  models.swift
-//  s2ar
-//
-//  Created by kokucy MacBookPro on 2018/12/20.
-//  Copyright © 2018年 Junya Ishihara. All rights reserved.
-//
-
-import Foundation
-
-class Models {
-    static let spaceship = [
-        //頭
-        [2, 0, -1, 2],
-        [2, 0, -1, 3, 85, 85, 85],
-        [2, 0, 0, 2],
-        [2, 0, 0, 3, 85, 85, 85],
-        //胴
-        [8, 0, -2, 0, 255, 51, 255],
-        [8, 1, -2, 0],
-        [8, 0, -2, 1],
-        [8, 1, -2, 1],
-        [8, 0, -1, 1, 255, 51, 255],
-        [8, 1, -1, 1],
-        [8, 1, -1, 0, 255, 51, 255],
-        //足
-        [2, 2, -1, -4, 170, 170, 170],
-        [2, 2, -1, -3, 170, 170, 170],
-        [2, 2, -1, -2, 170, 170, 170],
-        [2, 2, -1, -1, 170, 170, 170],
-        [2, 2, 0, -4, 170, 170, 170],
-        [2, 2, 0, -3, 170, 170, 170],
-        [2, 2, 0, -2, 170, 170, 170],
-        [2, 2, 0, -1, 170, 170, 170],
-        [2, 0, -3, -4, 170, 170, 170],
-        [2, 0, -3, -3, 170, 170, 170],
-        [2, 0, -3, -2, 170, 170, 170],
-        [2, 0, -3, -1, 170, 170, 170],
-        [2, 0, 2, -4, 170, 170, 170],
-        [2, 0, 2, -3, 170, 170, 170],
-        [2, 0, 2, -2, 170, 170, 170],
-        [2, 0, 2, -1, 170, 170, 170],
-    ]
-    
-    static let invader = [
-        //胴
-        [8, 0, -2, 0],
-        [8, 1, -2, 0],
-        [8, 0, -2, 1],
-        [8, 1, -2, 1],
-        [8, 0, -1, 1],
-        [8, 1, -1, 1],
-        [8, 1, -1, 0],
-        //足
-        [2, 1, -4, 0],
-        [2, 1, -3, 0],
-        [2, 1, -4, 1],
-        //手
-        [2, 2, -1, -1],
-        [2, 3, -1, -1],
-        [2, 3, 0, -1],
-        //角
-        [2, 1, 2, -1],
-        //顔（塗り）
-        [2, 0, -2, 1, 35, 35, 35],
-        [2, 1, 0, 1, 35, 35, 35],
-    ]
-    
-    static let human = [
-        //頭
-        [2, 0, 2, 0, 254, 204, 102],
-        [2, 0, 2, -1, 35, 35, 35],
-        [2, 0, 3, 0, 35, 35, 35],
-        [2, 0, 3, -1, 35, 35, 35],
-        //胴
-        [8, 0, -2, 0],
-        [8, 1, -2, 0],
-        [8, 0, -2, 1],
-        [8, 1, -2, 1],
-        [8, 0, -1, 1],
-        [8, 1, -1, 1],
-        [8, 1, -1, 0],
-        //足
-        [2, 1, -4, 0],
-        [2, 1, -3, 0, 254, 204, 102],
-        [2, 1, -4, 1],
-        //手
-        [2, 2, 1, 0],
-        [2, 2, 0, 0, 254, 204, 102],
-        [2, 2, 0, 0, 254, 204, 102],
-        [2, 2, -1, 0, 254, 204, 102],
-        [2, 2, -1, 1, 254, 204, 102],
-        [2, 2, -1, 2, 254, 204, 102],
-    ]
-    
-    static let cat = [
-        //胴
-        [4, 0, -3, 0],
-        [4, 1, -3, 0],
-        [4, 0, -3, 1],
-        [4, 1, -3, 1],
-        [4, 0, -2, 1],
-        [4, 1, -2, 1],
-        [4, 1, -2, 0],
-        [4, 0, -1, 1],
-        [4, 1, -1, 1],
-        [4, 1, -1, 0],
-        [4, 0, 0, 0],
-        [4, 1, 0, 0],
-        [4, 0, 0, 1],
-        [4, 1, 0, 1],
-        //足
-        [4, 1, -4, 1],
-        [4, 1, -4, 2],
-        //しっぽ
-        [2, 0, 0, -3],
-        [2, 0, 0, -4],
-        [2, 0, 1, -4],
-        //顔
-        [2, 0, -1, 2],
-        [2, 0, -1, 3],
-        [2, 1, -1, 2],
-        [2, 0, 0, 2],
-        [2, 0, 0, 3],
-        [2, 1, 0, 2, 35, 35, 35],
-        [2, 0, 1, 0],
-        [2, 1, 1, 0],
-        [2, 0, 1, 1],
-        [2, 1, 1, 1],
-        [2, 0, 1, 2],
-        [2, 1, 1, 2],
-        [2, 1, 2, 0],
-    ]
-    
-    static let dinosaur = [
-        //胴体１
-        [2, 0, -3, -2],
-        [2, 0, -3, -1],
-        [2, 0, -3, 0],
-        [2, 1, -3, -2],
-        [2, 1, -3, -1],
-        [2, 1, -3, 0],
-        //胴体2
-        [2, 0, -2, -2],
-        [2, 0, -2, -1],
-        [2, 0, -2, 0],
-        [2, 0, -2, 1],
-        [2, 1, -2, -2],
-        [2, 1, -2, -1],
-        [2, 1, -2, 0],
-        [2, 1, -2, 1],
-        //胴体3
-        [2, 0, -1, -1],
-        [2, 0, -1, 0],
-        [2, 0, -1, 1],
-        [2, 1, -1, -1],
-        [2, 1, -1, 0],
-        [2, 1, -1, 1],
-        [2, 1, -1, 2],
-        //胴体4
-        [2, 0, 0, -1],
-        [2, 0, 0, 0],
-        [2, 0, 0, 1],
-        [2, 1, 0, -1],
-        [2, 1, 0, 0],
-        [2, 1, 0, 1],
-        //胴体5
-        [2, 0, 1, 0],
-        [2, 0, 1, 1],
-        [2, 0, 1, 2],
-        [2, 0, 1, 3],
-        [2, 1, 1, 2],
-        [2, 1, 1, 1],
-        [2, 1, 1, 2],
-        [2, 1, 1, 3],
-        //胴体6
-        [2, 0, 2, 0],
-        [2, 0, 2, 1],
-        [2, 0, 2, 2],
-        [2, 0, 2, 3],
-        [2, 1, 2, 1],
-        [2, 1, 2, 2],
-        [2, 1, 2, 3],
-        //胴体7
-        [2, 0, 3, 1],
-        [2, 0, 3, 2],
-        [2, 1, 3, 1, 255, 255, 204],
-        [2, 1, 3, 2, 35, 35, 35],
-        //足
-        [2, 2, -4, 0],
-        [2, 2, -4, 1],
-        [2, 2, -3, 0],
-        [2, 2, -2, 0],
-        //しっぽ
-        [2, 0, -4, -4],
-        [2, 0, -4, -3],
-        [2, 0, -3, -3],
-    ]
-    
-    static let frog = [
-        //胴体1
-        [2, 0, -3, -3],
-        [2, 0, -3, -2],
-        [2, 0, -3, -1],
-        [2, 0, -3, 0],
-        [2, 0, -3, 1],
-        [2, 1, -3, -3],
-        [2, 1, -3, -2],
-        [2, 1, -3, -1],
-        [2, 1, -3, 0],
-        [2, 1, -3, 1],
-        //胴体2
-        [2, 0, -2, -3],
-        [2, 0, -2, 1],
-        [2, 1, -2, -3],
-        [2, 1, -2, -2],
-        [2, 1, -2, -1],
-        [2, 1, -2, 0],
-        [2, 1, -2, 1],
-        //胴体3
-        [2, 0, -1, -2],
-        [2, 0, -1, 1],
-        [2, 1, -1, -2],
-        [2, 1, -1, -1],
-        [2, 1, -1, 0],
-        [2, 1, -1, 1],
-        //胴体4
-        [2, 0, 0, -1],
-        [2, 0, 0, 1],
-        [2, 0, 0, 2],
-        [2, 1, 0, -1],
-        [2, 1, 0, 0],
-        [2, 1, 0, 1],
-        //胴体5
-        [2, 0, 1, 0],
-        [2, 0, 1, 1],
-        [2, 1, 1, 0],
-        [2, 1, 1, 1, 35, 35, 35],
-        //目
-        [2, 1, 2, 1, 255, 255, 204],
-        //前足
-        [2, 1, -4, 1],
-        [2, 1, -4, 2],
-        //後ろ足
-        [2, 3, -4, -3],
-        [2, 3, -4, -2],
-        [2, 3, -3, -3],
-        [2, 2, -3, -3],
-    ]
-    
-    static let tree = [
-        //幹
-        [4, 0, -4, 0, 85, 0 ,0],
-        [4, 0, -3, 0, 85, 0 ,0],
-        //枝１
-        [4, 3, -2, 0],
-        [4, 3, -2, 1],
-        [4, 2, -2, 1],
-        [4, 2, -2, 2],
-        [4, 1, -2, 2],
-        [4, 1, -2, 3],
-        [4, 0, -2, 3],
-        //枝2
-        [4, 2, -1, 0],
-        [4, 1, -1, 1],
-        [4, 0, -1, 2],
-        //枝3
-        [4, 1, 0, 0],
-        [4, 1, 0, 1],
-        [4, 0, 0, 1],
-        //枝4
-        [4, 1, 1, 0],
-        [4, 0, 1, 1],
-        //枝5
-        [4, 0, 2, 0],
-        [4, 0, 3, 0],
-    ]
-    
-    static let car = [
-        //タイヤ
-        [2, 2, -4, -4, 35, 35, 35],
-        [2, 2, -4, -3, 35, 35, 35],
-        [2, 2, -4, 0, 35, 35, 35],
-        [2, 2, -4, 1, 35, 35, 35],
-        //車体1
-        [2, 2, -3, -4, 35, 35, 35],
-        [2, 2, -3, -3, 35, 35, 35],
-        [2, 2, -3, -2],
-        [2, 2, -3, -1],
-        [2, 2, -3, 0, 35, 35, 35],
-        [2, 2, -3, 1, 35, 35, 35],
-        [2, 2, -3, 2],
-        [2, 2, -3, 3],
-        [2, 1, -3, -4],
-        [2, 1, -3, -3],
-        [2, 1, -3, -2],
-        [2, 1, -3, -1],
-        [2, 1, -3, 0],
-        [2, 1, -3, 1],
-        [2, 1, -3, 2],
-        [2, 1, -3, 3, 255, 51, 255],
-        [2, 0, -3, -4],
-        [2, 0, -3, -3],
-        [2, 0, -3, -2],
-        [2, 0, -3, -1],
-        [2, 0, -3, 0],
-        [2, 0, -3, 1],
-        [2, 0, -3, 2],
-        [2, 0, -3, 3, 255, 51, 255],
-        //車体２
-        [2, 2, -2, -4, 255, 255, 51],
-        [2, 2, -2, -3],
-        [2, 2, -2, -2],
-        [2, 2, -2, -1],
-        [2, 2, -2, 0],
-        [2, 2, -2, 1],
-        [2, 2, -2, 2],
-        [2, 2, -2, 3, 255, 255, 51],
-        [2, 1, -2, -4],
-        [2, 1, -2, 3],
-        [2, 0, -2, -4],
-        [2, 0, -2, 3],
-        //車体３
-        [2, 2, -1, -4],
-        [2, 2, -1, -3],
-        [2, 2, -1, -2],
-        [2, 2, -1, -1],
-        [2, 2, -1, 0],
-        [2, 2, -1, 1],
-        [2, 2, -1, 2],
-        [2, 2, -1, 3],
-        [2, 1, -1, -4],
-        [2, 1, -1, -3],
-        [2, 1, -1, -2],
-        [2, 1, -1, -1],
-        [2, 1, -1, 0],
-        [2, 1, -1, 1],
-        [2, 1, -1, 2],
-        [2, 1, -1, 3],
-        [2, 0, -1, -4],
-        [2, 0, -1, -3],
-        [2, 0, -1, -2],
-        [2, 0, -1, -1],
-        [2, 0, -1, 0],
-        [2, 0, -1, 1],
-        [2, 0, -1, 2],
-        [2, 0, -1, 3],
-        //車体4
-        [2, 2, 0, -4],
-        [2, 2, 0, -3, 255, 255, 204],
-        [2, 2, 0, -2, 255, 255, 204],
-        [2, 2, 0, -1, 255, 255, 204],
-        [2, 2, 0, 0, 255, 255, 204],
-        [2, 2, 0, 1],
-        [2, 2, 0, 2],
-        [2, 1, 0, -4, 255, 255, 204],
-        [2, 1, 0, 2, 255, 255, 204],
-        [2, 0, 0, -4, 255, 255, 204],
-        [2, 0, 0, 2, 255, 255, 204],
-        //車体5
-        [2, 2, 1, -4],
-        [2, 2, 1, -3, 255, 255, 204],
-        [2, 2, 1, -2, 255, 255, 204],
-        [2, 2, 1, -1, 255, 255, 204],
-        [2, 2, 1, 0, 255, 255, 204],
-        [2, 2, 1, 1],
-        [2, 1, 1, -4, 255, 255, 204],
-        [2, 1, 1, 1, 255, 255, 204],
-        [2, 0, 1, -4, 255, 255, 204],
-        [2, 0, 1, 1, 255, 255, 204],
-        //車体6
-        [2, 2, 2, -4],
-        [2, 2, 2, -3],
-        [2, 2, 2, -2],
-        [2, 2, 2, -1],
-        [2, 2, 2, 0],
-        [2, 1, 2, -4],
-        [2, 1, 2, -3],
-        [2, 1, 2, -2],
-        [2, 1, 2, -1],
-        [2, 1, 2, 0],
-        [2, 0, 2, -4],
-        [2, 0, 2, -3],
-        [2, 0, 2, -2],
-        [2, 0, 2, -1],
-        [2, 0, 2, 0],
-    ]
-    
-    static let airplane = [
-        //胴体１
-        [2, 0, -4, -4],
-        [2, 0, -4, -3],
-        [2, 0, -4, -2],
-        [2, 0, -4, -1],
-        [2, 0, -4, 0],
-        [2, 0, -4, 1],
-        [2, 0, -4, 2],
-        [2, 0, -4, 3],
-        //胴体２
-        [2, 0, -3, -4],
-        [2, 0, -3, -3],
-        [2, 0, -3, -2],
-        [2, 0, -3, -1],
-        [2, 0, -3, 0],
-        [2, 0, -3, 1],
-        [2, 0, -3, 2],
-        [2, 0, -3, 3],
-        //胴体３
-        [2, 0, -2, -4],
-        [2, 0, -2, -3],
-        [2, 0, -2, -2],
-        [2, 0, -2, -1],
-        [2, 0, -2, 0],
-        [2, 0, -2, 1],
-        [2, 0, -2, 2, 255, 255, 204],
-        //垂直尾翼
-        [2, 0, -1, -4],
-        [2, 0, -1, -3],
-        [2, 0, 0, -4],
-        //水平尾翼
-        [2, 1, -3, -4],
-        [2, 1, -3, -3],
-        [2, 2, -3, -4, 255, 255, 204],
-        [2, 2, -3, -3, 255, 255, 204],
-        //主翼
-        [2, 1, -3, 0],
-        [2, 1, -3, 1],
-        [2, 2, -3, 0],
-        [2, 2, -3, 1],
-        [2, 3, -3, 0, 255, 255, 204],
-        [2, 3, -3, 1, 255, 255, 204],
-    ]
-    
-    static let sword = [
-        [2, 0, -1, -4],
-        [2, 0, -1, -3],
-        [2, 0, -1, -2],
-        [2, 0, -1, -1, 255, 255, 204],
-        [2, 0, -1, 0, 255, 255, 204],
-        [2, 0, -1, 1, 255, 255, 204],
-        [2, 0, -1, 2, 255, 255, 204],
-        [2, 0, -1, 3, 255, 255, 204],
-        [2, 0, 0, -4],
-        [2, 0, 0, -3],
-        [2, 0, 0, -2],
-        [2, 0, 0, -1, 255, 255, 204],
-        [2, 0, 0, 0, 255, 255, 204],
-        [2, 0, 0, 1, 255, 255, 204],
-        [2, 0, 0, 2, 255, 255, 204],
-        [2, 0, 0, 3, 255, 255, 204],
-        [2, 1, -1, -2, 0, 0, 68],
-        [2, 0, -2, -2, 0, 0, 68],
-        [2, 1, 0, -2, 0, 0, 68],
-        [2, 0, 1, -2, 0, 0, 68],
-    ]
-    
-    static let fire = [
-        [2, 0, -2, -4, 255, 255, 51],
-        [2, 0, -2, -3, 255, 255, 51],
-        [2, 0, -2, -2, 255, 255, 51],
-        [2, 0, -2, -1, 255, 255, 51],
-        [2, 0, -2, 0],
-        [2, 1, -2, -4],
-        [2, 1, -2, -3],
-        [2, 1, -2, -2],
-        [2, 1, -2, -1],
-        [2, 1, -2, 0],
-        [2, 1, -1, -4, 255, 255, 51],
-        [2, 1, -1, -3, 255, 255, 51],
-        [2, 1, -1, -2, 255, 255, 51],
-        [2, 1, -1, -1, 255, 255, 51],
-        [2, 1, -1, 0],
-        [2, 1, 0, -4, 255, 255, 51],
-        [2, 1, 0, -3, 255, 255, 51],
-        [2, 1, 0, -2, 255, 255, 51],
-        [2, 1, 0, -1, 255, 255, 51],
-        [2, 1, 0, 0],
-        [2, 0, 1, -4, 255, 255, 51],
-        [2, 0, 1, -3, 255, 255, 51],
-        [2, 0, 1, -2, 255, 255, 51],
-        [2, 0, 1, -1, 255, 255, 51],
-        [2, 0, 1, 0],
-        [2, 1, 1, -4],
-        [2, 1, 1, -3],
-        [2, 1, 1, -2],
-        [2, 1, 1, -1],
-        [2, 1, 1, 0],
-        [2, 0, -1, 1],
-        [2, 0, -1, 2],
-        [2, 0, -1, 3],
-        [2, 0, 0, 1],
-        [2, 0, 0, 2],
-        [2, 0, 0, 3],
-    ]
-    
-    static let explosion = [
-        //random points
-        [8, 0, 0, 0],
-        [1, -1, 0 ,-4],
-        [1, -1, 3, -3],
-        [2, 3, -1, 0],
-        [1, 3, 0, -2],
-        [1, -3, 1, 3],
-        [1, -4, 2, 0],
-        [1, -4, -2, 1],
-        [4, 0, -4, 2],
-        [1, 1, 2, 2],
-        [1, 1, -4, 3],
-        [2, -1, -2, -4],
-        [1, 0, -4, -2],
-        [1, -1, 0, 1],
-        [1, 1, -1, -3],
-        [4, -4, 0, -1],
-        [1, 0, -4, -2],
-    ]
-    
-    static let box1 = [
-        [1, 0, 0, 0],
-    ]
-    
-    static let box2 = [
-        [8, 0, -1, 0],
-    ]
-    
-    static let box4 = [
-        [8, 0, -2, 0],
-        [8, 1, -2, 0],
-        [8, 0, -2, 1],
-        [8, 1, -2, 1],
-        [8, 1, -1, 0],
-        [8, 0, -1, 1],
-        [8, 1, -1, 1],
-    ]
-    
-    static let box8 = [
-        [8, 0, -4, 0],
-        [8, 1, -4, 0],
-        [8, 2, -4, 0],
-        [8, 3, -4, 0],
-        [8, 0, -4, 1],
-        [8, 1, -4, 1],
-        [8, 2, -4, 1],
-        [8, 3, -4, 1],
-        [8, 0, -4, 2],
-        [8, 1, -4, 2],
-        [8, 2, -4, 2],
-        [8, 3, -4, 2],
-        [8, 0, -4, 3],
-        [8, 1, -4, 3],
-        [8, 2, -4, 3],
-        [8, 3, -4, 3],
-        [8, 0, -3, 3],
-        [8, 1, -3, 3],
-        [8, 2, -3, 3],
-        [8, 3, -3, 3],
-        [8, 0, -2, 3],
-        [8, 1, -2, 3],
-        [8, 2, -2, 3],
-        [8, 3, -2, 3],
-        [8, 0, -1, 3],
-        [8, 1, -1, 3],
-        [8, 2, -1, 3],
-        [8, 3, -1, 3],
-        [8, 3, -3, 0],
-        [8, 3, -3, 1],
-        [8, 3, -3, 2],
-        [8, 3, -2, 0],
-        [8, 3, -2, 1],
-        [8, 3, -2, 2],
-        [8, 3, -1, 0],
-        [8, 3, -1, 1],
-        [8, 3, -1, 2]
-    ]
-    
-    static let default_model = [
-        [4, 0 ,0, 0]
-    ]
-}
